+++ packages/visual-editor/src/components/configs/index.ts
@@ -1,0 +1,6 @@
+export { mainConfig, type MainConfigProps } from "./mainConfig.tsx";
+export {
+  directoryConfig,
+  type DirectoryConfigProps,
+} from "./directoryConfig.tsx";
+export { locatorConfig, type LocatorConfigProps } from "./locatorConfig.tsx";