+++ packages/visual-editor/src/internal/components/ThemeEditor.tsx
@@ -95,13 +95,13 @@ export const ThemeEditor = (props: ThemeEditorProps) => {
           // strip ui state
           return { id: h.id, state: { data: { ...h.state.data } } };
         });
+        // @ts-expect-error https://github.com/measuredco/puck/issues/673
         setPuckInitialHistory({
-          // @ts-expect-error https://github.com/measuredco/puck/issues/673
           histories: histories,
           index: localHistoryIndex,
           appendData: false,
         });
-        const layoutToSend = histories[histories.length - 1].state.data;
+        const layoutToSend = histories[histories.length - 1]?.state?.data;
         sendDevLayoutSaveStateData({
           payload: { devSaveStateData: JSON.stringify(layoutToSend) },
         });