+++ packages/visual-editor/src/utils/applyTheme.ts
@@ -23,6 +23,7 @@ export type StreamDocument = {
     name?: string;
     visualEditorConfig?: string;
     isPrimaryLocale?: boolean;
+    entityPageSetUrlTemplates?: string;
   };
 };
 
@@ -121,7 +122,7 @@ export const updateThemeInEditor = async (
       PUCK_PREVIEW_IFRAME_ID
     ) as HTMLIFrameElement;
     const pagePreviewStyleTag =
-      iframe.contentDocument?.getElementById(THEME_STYLE_TAG_ID);
+      iframe?.contentDocument?.getElementById(THEME_STYLE_TAG_ID);
     if (pagePreviewStyleTag) {
       observer.disconnect();
       pagePreviewStyleTag.innerText = newThemeTag;