+++ packages/visual-editor/src/utils/index.ts
@@ -6,7 +6,7 @@ export { applyAnalytics, getAnalyticsScopeHash } from "./applyAnalytics.ts";
 export { applyHeaderScript } from "./applyHeaderScript.ts";
 export { applyTheme, type StreamDocument } from "./applyTheme.ts";
 export { themeManagerCn, themeManagerTwMergeConfiguration } from "./cn.ts";
-export { filterComponentsFromConfig } from "./filterComponents.tsx";
+export { filterComponentsFromConfig } from "./filterComponents.ts";
 export { getPageMetadata } from "./getPageMetadata.ts";
 export {
   migrate,
@@ -31,6 +31,7 @@ export {
   ThemeOptions,
   type BackgroundStyle,
   type HeadingLevel,
+  VisualEditorComponentsContentPath,
 } from "./themeConfigOptions.ts";
 export { type ThemeConfig, themeResolver } from "./themeResolver.ts";
 export { VisualEditorProvider } from "./VisualEditorProvider.tsx";