+++ packages/visual-editor/src/utils/resolveUrlTemplate.test.ts
@@ -27,6 +27,27 @@ const mockStreamDocument: StreamDocument = {
   }),
 };
 
+const mockDirectoryDocument: StreamDocument = {
+  name: "Yext",
+  id: "123",
+  locale: "en",
+  address: {
+    line1: "61 9th Ave",
+    city: "New York",
+    region: "NY",
+    country: "USA",
+  },
+  __: {
+    isPrimaryLocale: true,
+    codeTemplate: "directory",
+    entityPageSetUrlTemplates: JSON.stringify({
+      primary: "[[address.region]]/page/[[id]]",
+      alternate: "[[locale]]/[[address.region]]/page/[[id]]",
+    }),
+  },
+  _pageset: JSON.stringify({}),
+};
+
 describe("resolveUrlTemplate", () => {
   it("resolves primary template for primary locale", () => {
     const result = resolveUrlTemplate(mockStreamDocument, "");
@@ -47,7 +68,9 @@ describe("resolveUrlTemplate", () => {
     const alternateLocaleDoc = {
       ...mockStreamDocument,
       locale: "Zh_HANS-hk",
-      __: {},
+      __: {
+        isPrimaryLocale: false,
+      },
     };
 
     assert.equal(
@@ -56,11 +79,11 @@ describe("resolveUrlTemplate", () => {
     );
   });
 
-  it("defaults to alternate template if '__' is missing", () => {
+  it("defaults to primary template if '__' is missing", () => {
     // eslint-disable-next-line no-unused-vars
     const { __, ...docWithoutPrimaryInfo } = mockStreamDocument;
     const result = resolveUrlTemplate(docWithoutPrimaryInfo, "");
-    assert.equal(result, "en/ny/new-york/61-9th-ave");
+    assert.equal(result, "ny/new-york/61-9th-ave");
   });
 
   it("gracefully resolves empty fields", () => {
@@ -136,7 +159,7 @@ describe("resolveUrlTemplate", () => {
     assert.equal(result, "../ny/new-york/61-9th-ave");
   });
 
-  it("use fallback if primary template is missing for primary locale", () => {
+  it("use alternate if primary template is missing for primary locale", () => {
     const docWithoutPrimaryTemplate = {
       ...mockStreamDocument,
       _pageset: JSON.stringify({
@@ -148,10 +171,10 @@ describe("resolveUrlTemplate", () => {
       }),
     };
     const result = resolveUrlTemplate(docWithoutPrimaryTemplate, "../");
-    assert.equal(result, "../ny/new-york/61-9th-ave");
+    assert.equal(result, "../en/ny/new-york");
   });
 
-  it("use fallback if alternate template is missing for alternate locale", () => {
+  it("use primary template if alternate template is missing for alternate locale", () => {
     const docWithoutAlternateTemplate = {
       ...mockStreamDocument,
       __: { isPrimaryLocale: false },
@@ -165,7 +188,41 @@ describe("resolveUrlTemplate", () => {
       }),
     };
     const result = resolveUrlTemplate(docWithoutAlternateTemplate, "../");
+    assert.equal(result, "../ny/new-york/61-9th-ave");
+  });
+
+  it("use fallback if all templates are missing", () => {
+    const docWithoutAlternateTemplate = {
+      ...mockStreamDocument,
+      __: { isPrimaryLocale: false },
+      locale: "es",
+      _pageset: JSON.stringify({
+        config: {
+          urlTemplate: {},
+        },
+      }),
+    };
+    const result = resolveUrlTemplate(docWithoutAlternateTemplate, "../");
     assert.equal(result, "../es/ny/new-york/61-9th-ave");
+  });
+
+  it("use primary if isPrimaryLocale field is missing", () => {
+    const docWithoutAlternateTemplate = {
+      ...mockStreamDocument,
+      __: {},
+      locale: "fr",
+      _pageset: JSON.stringify({
+        config: {
+          urlTemplate: {
+            primary: "[[address.region]]/[[address.city]]/[[address.line1]]",
+            alternate:
+              "[[locale]]/[[address.region]]/[[address.city]]/[[address.line1]]",
+          },
+        },
+      }),
+    };
+    const result = resolveUrlTemplate(docWithoutAlternateTemplate, "../");
+    assert.equal(result, "../ny/new-york/61-9th-ave");
   });
 
   it("use alternateFunction when provided to resolve URL template", () => {
@@ -197,5 +254,22 @@ describe("resolveUrlTemplate", () => {
       __: { isPrimaryLocale: false },
     };
     expect(() => resolveUrlTemplate(alternateLocaleDoc, "")).toThrowError();
+  });
+
+  it("handles primary url template on directory pages", () => {
+    expect(resolveUrlTemplate(mockDirectoryDocument, "")).toBe("ny/page/123");
+  });
+
+  it("handles alternate url templates on directory pages", () => {
+    expect(
+      resolveUrlTemplate(
+        {
+          ...mockDirectoryDocument,
+          __: { ...mockDirectoryDocument.__, isPrimaryLocale: false },
+          locale: "es",
+        },
+        ""
+      )
+    ).toBe("es/ny/page/123");
   });
 });